Modify set state to use a transformation function
In general, because setting state in ReactJS is asynchronous, there is no
guarantee that our last call to set state has completed when our next call to
set state occurs. For this reason, it is preferable to pass a
transformation function into our call to set state - rather than pass an
explicity value derived from the current props.